Options and identifier symbols

+++ src/PDACleaner.cc
@@ -388,7 +388,7 @@ static NFA * build_NFA(PDA *P_0, PDATransitionSet *U_1, Sqs *sqs_map) {
     BMap *b_map = new BMap();
     
     // Initial state of the NFA
-    GeneratedState m_0("-4");
+    GeneratedState m_0("m_0");
     
     // Stateset of the NFA
     StateSet Q;
@@ -432,12 +432,9 @@ static NFA * build_NFA(PDA *P_0, PDATransitionSet *U_1, Sqs *sqs_map) {
 
 static PDA * simplify_PDA(PDA *P) {
     Bottom b_0;
-//    GeneratedState q_0_hat("q_0_hat");
-//    GeneratedState q_e("q_e");
-//    GeneratedState q_f("q_f");
-    GeneratedState q_0_hat("-1");
-    GeneratedState q_e("-2");
-    GeneratedState q_f("-3");
+    GeneratedState q_0_hat("q_0_hat");
+    GeneratedState q_e("q_e");
+    GeneratedState q_f("q_f");
     SymbolString Z;
     Z.append(&b_0);
     StateSet F;
@@ -456,28 +453,28 @@ static PDA * simplify_PDA(PDA *P) {
     // Add necessary transitions to Delta
     SymbolString empty_string;
     
-    PDATransition t(&q_0_hat, &empty_string, 0, P->get_q_0(), P->get_Z(), START_UP);
+    PDATransition t(&q_0_hat, &empty_string, "", P->get_q_0(), P->get_Z(), START_UP);
     P_0->get_Delta()->add(&t);
     
     for (auto it = P->get_F()->begin(); it != P->get_F()->end(); it++) {
-        PDATransition t(*it, &empty_string, 0, &q_e, &empty_string, CLOSE_DOWN);
+        PDATransition t(*it, &empty_string, "", &q_e, &empty_string, CLOSE_DOWN);
         P_0->get_Delta()->add(&t);
     }
 
     for (auto it = P->get_Gamma()->begin(); it != P->get_Gamma()->end(); it++) {
         SymbolString pop;
         pop.append(*it);
-        PDATransition t(&q_e, &pop, 0, &q_e, &empty_string, CLOSE_DOWN);
+        PDATransition t(&q_e, &pop, "", &q_e, &empty_string, CLOSE_DOWN);
         P_0->get_Delta()->add(&t);
     }
 
-    PDATransition r(&q_e, &Z, 0, &q_f, &empty_string, CLOSE_DOWN);
+    PDATransition r(&q_e, &Z, "", &q_f, &empty_string, CLOSE_DOWN);
     P_0->get_Delta()->add(&r);
 
     return P_0;
 }
 
-PDACleanerResult clean_PDA(PDA *P) {
+PDACleanerResult clean_PDA(PDA *P, bool clean) {
     Sqs *sqs_map = new Sqs();
     PDA *P_0 = simplify_PDA(P);
 
@@ -505,15 +502,19 @@ PDACleanerResult clean_PDA(PDA *P) {
     remove_extensions(U_1);
     remove_extensions(U_2);
 
-    // Create a clean PDA
-    PDA *P_clean = P->clone();
-    P_clean->get_Delta()->remove_all(U_1);
-    P_clean->get_Delta()->remove_all(U_2);
+    PDACleanerResult result = { U_1, U_2, nullptr };
+    
+    // Create a clean PDA, if requested
+    if (clean) {
+        PDA *P_clean = P->clone();
+        P_clean->get_Delta()->remove_all(U_1);
+        P_clean->get_Delta()->remove_all(U_2);
+        
+        // Remove unused states
+        remove_unused_states(P_clean);
 
-    // Remove unused states
-    remove_unused_states(P_clean);
-
-    PDACleanerResult result = { U_1, U_2, P_clean };
+        result.P_clean = P_clean;
+    }
 
     delete sqs_map;
     delete N;
